Removes version 2 of updated luxon files
