+++ packages/palette/package.json
@@ -116,8 +116,6 @@
     "d3-interpolate": "^1.3.2",
     "d3-shape": "^1.3.5",
     "luxon": "^1.15.0",
-    "moment": "^2.23.0",
-    "moment-timezone": "^0.5.25",
     "rc-slider": "^8.6.2",
     "react-lazy-load-image-component": "^1.3.2",
     "react-live": "^1.12.0",
