+++ packages/palette/src/elements/TimeRemaining/TimeRemaining.tsx
@@ -1,7 +1,6 @@
-import moment from "moment"
+import { DateTime, Duration } from "luxon"
 import React from "react"
-
-import { Sans } from "../"
+import { Sans } from ".."
 import { color } from "../../helpers"
 
 const pad = (n: number) => n.toString().padStart(2, "0")
@@ -12,20 +11,26 @@ export const TimeRemaining: React.SFC<{
   countdownEnd: string
   highlight: Parameters<typeof color>[0]
 }> = ({ currentTime, countdownEnd, highlight = "purple100" }) => {
-  const timeRemaining = moment.duration(
-    moment(countdownEnd).diff(moment(currentTime))
+  const timeRemaining = Duration.fromISO(
+    DateTime.fromISO(countdownEnd)
+      .diff(DateTime.fromISO(currentTime))
+      .toString()
   )
+  const days = Math.floor(timeRemaining.as("days"))
+  const hours = Math.floor(timeRemaining.as("hours") % 24)
+  const minutes = Math.floor(timeRemaining.as("minutes") % 60)
+  const seconds = Math.floor(timeRemaining.as("seconds") % 60)
 
   return (
     <Sans size="3" color={highlight} weight="medium">
-      {Math.floor(timeRemaining.asSeconds()) <= 0 ? (
+      {Math.floor(timeRemaining.seconds) <= 0 ? (
         "0 days"
       ) : (
         <>
-          {timeRemaining.days() > 0 && pad(timeRemaining.days()) + "d "}
-          {timeRemaining.hours() > 0 && pad(timeRemaining.hours()) + "h "}
-          {timeRemaining.minutes() > 0 && pad(timeRemaining.minutes()) + "m "}
-          {pad(timeRemaining.seconds()) + "s"}
+          {days > 0 && pad(days) + "d "}
+          {hours > 0 && pad(hours) + "h "}
+          {minutes > 0 && pad(minutes) + "m "}
+          {pad(seconds) + "s"}
         </>
       )}
       <span> left</span>
